Removed variable declaration playButton
It is unnecessary to declare a variable that stores a reference to an
HTML element if it is only used once. The playButton variable was
referred to only once: to set the display property of the play button
to none. Thus the variable was removed in favor for calling the
getElementById method together with setting the display property, in
a single statement.