Replaced links with buttons
The anchor tags were replaced with button elements. The showHelp
function is called from the button with the onclick property, from
inside the startGame function. Buttons are more suitable in this case
since they are displaying (or toggling) pop-ups, not redirecting to
another page. Using the onclick property rather than a link with
href:javascript is more consistent with the unobtrusive javascript
approach.