Added help access before game starts
The player should be able to view the help, about, and options before
initiating the game.

To add this functionality, the onclick property of the help button
which fires the showHelp function was moved from inside the startGame
function into global scope. showHelp was also moved from startGame into
global scope.

This is a quick-and-dirty implementation. This should probably be
changed for better methods such as self executing functions, AMD, module
pattern, namespaces, closures, etc.