Fixed multiple help dialogue creation
The showHelp function fired by the help button was creating multiple
help divs in the DOM instead of only creating one and hiding and
displaying it. Since the expression of the if statement was testing if
the variable help is undefined, control never followed the else
statement because each time the showHelp function is called, the help
variable is created anew. (ie local variables as they pertain to
function scope)

This was fixed by checking instead for the creation of the help div in
the DOM using the getElementById method.