Fixed tokens covering up help dialogue
When the help button is clicked before the play button, the
help div is added before the token elements on the DOM list, which
causes the tokens to cover up the help dialogue when the help button is
clicked later during the game.

This was fixed by specifying the z-index properties of the token and
help classes to 0 and 1 respectively.