Removed about button
Having two buttons with a question mark and the letter 'i' is somewhat
ambiguous, so the about 'i' button has been removed. This allows for a
simpler interface with only help and options buttons. The information
that was meant to be in the about menu can be added to the help menu.