Moved the help menu as a child of play-area
All the HTML elements that are a part of the game should be added as
children of play-area to contain and separate them from other web page
elements when the game is deployed on a website.

The change introduces a graphical layering error similar to the one
resolved in commit a215, where the tokens are displayed above the help
menu. This is most likely caused because the help menu and the tokens no
longer share the same parents, thus the z-index properties between the
two no longer apply.

Also added dark gray background color to body of nim.html to reduce eye
strain for development purposes.