Moved tokens as children of play-area
All token objects are structured in the DOM as children of the play-area
div.

The help menu display error from the previous commit has been fixed. It
was caused by two different issues: 1) The help menu shape would appear
distorted after closing and reopening it, and 2) The tokens would cover
up the help menu. The first issue was fixed by using the visibility
property instead of display. The second was fixed by setting the
position property of the help menu to relative, since the position
must be set to absolute, relative, or fixed for the z-index to work.