Restructured code using the module pattern
Changed play-button from a link to a button element. Game starts by
calling the init() public method fired from play-button's onclick
attribute.

Had to move definition of numOfHeaps from the module into startGame()
since it was throwing an undefined error for getRandomInt(). This is
probably because functions do not hoist when declared inside a child
block.