Added remove() method for Token object
There is no need to have access to the tokens array since "this" refers
to the object itself.

At this point, a Token object is only capable of removing its own
element from the DOM, not any other neighboring Token elements.

May need to assign null to the Token object so that the garbage
collector can release it from memory, or use delete operator on the
Token's element property.