Renamed some methods as functions
Made distinction between functions and methods to clarify their context.
If it is associated with an object, like in Token, then it is referred
to as a method since it represents a capability of that object. If it
is not associated with an object, then it is referred to as a function.

This is just to give further semantic meaning since technically all the
functions are methods of the Nim module.