Added computer's turn
As with the event listeners, it was necessary to correct the value of
"this" with bind() when calling the remove method in setTimeout() or
else remove() would fail with undefined errors.

An alternate solution is to call remove() inside of an anonymous
function when using setTimeout(), but bind() is more efficient.